+++ ios/CHANGELOG.md
@@ -22,14 +22,13 @@ Line wrap the file at 100 chars.                                              Th
 * **Security**: in case of vulnerabilities.
 
 ## Unreleased
+### Fixed
+- Broken DAITA settings view on iOS 15.
 
 ## [2025.1 - 2025-01-14]
 ### Added
 - Update to DAITA v2 - now machines are provided by relays dynamically instead
   of using bundled ones.
-
-### Fixed
-- Fix DAITA settings view on iOS 15.
 
 ## [2024.11 - 2024-12-12]
 ### Added